+++ src/Authentication/Contracts/WebAuthnInterface.php
@@ -1,5 +1,16 @@
 <?php
 
+/**
+ * WebAuthnInterface contract.
+ *
+ * @package    ArtisanPack_UI
+ * @subpackage SecurityAdvancedAuth
+ *
+ * @author     Jacob Martella <support@artisanpackui.dev>
+ *
+ * @since      1.0.0
+ */
+
 declare( strict_types=1 );
 
 namespace ArtisanPackUI\SecurityAdvancedAuth\Authentication\Contracts;
@@ -83,5 +94,5 @@ interface WebAuthnInterface
     /**
      * Invalidate a challenge after use.
      */
-    public function invalidateChallenge( string $challenge): void;
+    public function invalidateChallenge( string $challenge ): void;
 }